Update and rename parsecwopinsol.R to parsecwopinsolc.R
the original file confused zenith angle and elevation angle.  It did not matter to the documented variables, only to cwopR, which is supposed to be the radius of an equatorial earth ring that would occult the sun for a given day, latitude, longitude.  The expression doesn't take time of day into account, but this matters little except around dawn and dusk.