Create ams2016.Rpres
this is a draft of the markdown file that creates most of the figures in the AMS 2016 presentation.  In the event we couldn't use the HTML5 output directly; it had to be cut and pasted into a PPT!!  Still, this markdown shows how to prepare most slides of the presentation directly from the database.